Store_AI_agent_cum_chatbot/requirements.txt
sentence-transformers
numpy
requests
pandas
python-dotenv
